==> Notebooks/6. Reference & Cleaning.ipynb <==
import numpy as np
import pandas as pd
from time import time
import re
import random
import pickle
from datetime import datetime, timedelta

df = pd.read_csv('D:/Capstone/dataset/emails_total_preprocessed_raw.csv')


df

df['domain']=df['from'].str.extract(r"@(.*)")

def email_clean(domain, email):
    output=[]
    for i in range(len(domain)):
        if domain[i] =='enron.com':
            output.append(email[i])
        else:
            output.append(domain[i])
    return output

df['from_clean'] = email_clean(df['domain'],df['from'])

user_list=df['user'].unique().tolist()
email_list=df['from_clean'].unique().tolist()

df_employee = pd.DataFrame({'uuid':[i for i in range(len(user_list))],'user':user_list})

df_email = pd.DataFrame({'eid':[i for i in range(len(email_list))],'from_clean':email_list})


df_messages = df[['Unnamed: 0','from_clean','user','type','subject','body','date', 'bcc']]
df_messages.columns = ['mid','from_clean','user','type','subject','body','date', 'bcc']

#email_counts=df['from'].value_counts().rename('email_counts')
#df_email=df.merge(email_counts.to_frame(),
#             left_on='from',
#             right_index=True)

df_messages.shape

def make_df(df,cnt_what,by2_criteria,by2,by1,by_list):
    new_df= pd.DataFrame({})
    for i in by_list:
        a=df[(df[by1]==i) & (df[by2]==by2_criteria)][cnt_what].value_counts()
        b=a.to_frame()
        b['email'] =b.index
        b['user']=i
        new_df= pd.concat([new_df,b])
    return new_df

df_alias=make_df(df,'from_clean','sent','type','user',user_list)
df_alias.reset_index(drop=True, inplace=True)
df_alias.columns=['sent_no','email','user']

#verify if we should keep aliases that have less than 10 emails sent. (did they save received emails to 'sent' folder)
check=df_alias[df_alias.sent_no<10]
len(check)
#check.reset_index(drop=True, inplace=True)

#check.loc[:80]

#alias 'vkaminski@aol.com' would be lost after domain appending. 
#keep
keep_list=['l..gay@enron.com','michele.lokay@enron.com',
           'stacey.white@enron.com','paul.thomas@enron.com',
           'eric.linder@enron.com','steven.merris@enron.com']

df_alias.head()

df1 = df_alias[df_alias['sent_no']>9]
for i in keep_list: 
    df1 = pd.concat([df1, df_alias[df_alias['email']==i]])

df_alias=df1

check=df_alias[df_alias.sent_no<10]
df_alias.reset_index(drop=True, inplace=True)

display(df_alias.head())
display(df_alias.tail())

for i in keep_list:
    display(df_alias[df_alias.email == i])

#df_email = df_email.merge(df_employee, 
#                          on ='user')
#df_email.drop(columns='user',inplace=True)
#df_messages.columns = ['eid','email','type','subject','body','date', 'bcc']

print('number of aliases:', df_alias.shape[0])
df_alias=df_alias.drop(columns ='sent_no')
df_alias.columns =['from_clean','user']

df_alias = df_alias.append(pd.Series(['sharris1@enron.com','harris-s'], index=df_alias.columns ), ignore_index=True)
df_alias = df_alias.append(pd.Series(['steven.harris@enron.com','harris-s'], index=df_alias.columns ), ignore_index=True)

keep_list

df_alias.head()

len(df_messages)

df_alias = df_alias.merge(df_email, on='from_clean')
df_alias = df_alias.merge(df_employee,on ='user')

df_alias[df_alias.duplicated('eid')]

df_alias[df_alias['eid']==1831]

df_alias[df_alias['eid']==103]

df_alias[df_alias['eid']==234]

df_email.to_csv('D:/Capstone/dataset/database/df_email.csv')

df_alias.to_csv('D:/Capstone/dataset/database/alias.csv')

df_messages=df_messages.merge(df_employee, on='user')
df_messages=df_messages.merge(df_email, on='from_clean')

df_messages.info()

len(df_messages)

#fix emails that are not captured in sent folders.
for i in range(len(df)):
    user = df_messages['uuid'][i]
    user_alias = df_alias[df_alias['uuid']==user]['eid'].tolist()
    if (df_messages['eid'][i] in user_alias) & (df_messages['type'][i] == 'received'):
        df_messages['type'][i] = 'sent'

df_messages.tail()

dates=df_messages['date'].tolist()
times=[i[-9:-2] for i in dates]

df_messages['times']=times

df_messages.info()

#infer date
df_messages['date'] = pd.to_datetime(df_messages['date'],infer_datetime_format = True)

df_messages.body[1337]

df_messages['unique']= df_messages['subject'].map(str)+ '-' +df_messages['from_clean'].map(str) + '-' + df_messages['date'].map(str) + '-' + df_messages['times'].map(str)

df_messages.unique.unique()[138168]

df_messages[df_messages.unique=='Memo on Mitigation Measures -sarah.novosel@enron.com-2000-08-04 02:42:00-02:42:0']

mid, unique = pd.factorize(df_messages.unique.tolist(),sort =True)

df_messages['mid']=mid

df_messages[df_messages.unique==' Master Agreement Structural Issues Working Group Meeting Wed nesd ay March at a.m. NY time p.m. London time -isda-eur.org-2001-03-20 09:08:00-09:08:0']

df_messages.drop(columns = ['times','unique'],inplace=True)

df_messages.to_csv('D:/Capstone/dataset/database/messages.csv')
