==> Notebooks/7. Merging.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime, timedelta

df= pd.read_csv('D:/Capstone/dataset/database/messages.csv')
df1= pd.read_csv('D:/Capstone/dataset/topic_matrix_final.csv')

df= df.merge(df1, on ='Unnamed: 0')
#infer date
df['date'] = pd.to_datetime(df['date'],infer_datetime_format = True)

df.columns

df.drop(columns =['Unnamed: 0','topic_0', 'topic_1', 'topic_2',
       'topic_3', 'topic_4', 'topic_5', 'topic_6', 'pred_topic', 'max_prob'], inplace=True)

#renamed columns
df.columns=['mid','address','user','type','subject','body','date','bcc','uuid','eid','topic']

topic_dict={'topic_0':'office communications',
                'topic_1':'industry expertise',
                'topic_2':'social',
                'topic_3':'operational compliance',
                'topic_4':'trading',
                'topic_5':'legal',
                'topic_6':'spam & html communique',
                'unclassified':'unclassified'
               }

df.head()

topic_list=[]
for i in range(len(df)):
        t= df['topic'][i]
        h = topic_dict.get(t)
        topic_list.append(h)

df['topic_h']=topic_list

df.info()

df.to_pickle('D:/Capstone/dataset/database/messages_combined.pkl')
df.to_csv('D:/Capstone/dataset/database/messages_combined.csv')

df.head(2)

df.shape

#drop unclassified rows
df=df[df.topic != 'unclassified']

print(df.shape)
print(144785/473749)

df.to_pickle('D:/Capstone/dataset/database/messages_combined_reduced.pkl')
df.to_csv('D:/Capstone/dataset/database/messages_combined_reduced.csv')

len(df)
