==> Notebooks/9. N4J.ipynb <==
import pandas as pd
import numpy as np

dfr= pd.read_pickle('D:/Capstone/dataset/database/messages_combined_reduced.pkl')
df= pd.read_pickle('D:/Capstone/dataset/database/messages_combined.pkl')
alias= pd.read_csv('D:/Capstone/dataset/database/N4J/alias.csv')

print(len(dfr))
print(len(df))

user_dict=dict(zip(alias['uuid'].unique().tolist(), alias['user'].unique()))
alias.drop(columns = ['Unnamed: 0'],inplace=True)
a_list=alias['eid'].unique().tolist()
print(len(a_list))

dfr_a=pd.DataFrame()
for a in a_list:
    result = dfr[(dfr['eid']==a)]
    dfr_a=dfr_a.append(result)

dfr_a.shape[0]

dfr_a.to_csv('D:/Capstone/dataset/database/N4J/internal.csv')

dfr does not have 'unclassified' topics, we would want to include these only for intra-network analysis.
Lets extract this from df where it includes all topics, including 'unclassified' ones.
Let us also include this into the reduced dataframe. 

dfr_rec=dfr[dfr.type=='received']
df_rec =df[df.type=='received']

dfr_rec_a= pd.DataFrame()
df_rec_a= pd.DataFrame()

for a in a_list:
    result = dfr_rec[(dfr_rec['eid']==a)]
    dfr_rec_a=dfr_rec_a.append(result)

for a in a_list:
    result = df[(df['eid']==a)]
    df_rec_a=df_rec_a.append(result)

print(len(dfr_rec_a))
print(len(df_rec_a))
print(len(dfr_rec_a.eid.unique()))
print(len(df_rec_a.eid.unique()))

print(len(dfr_rec_a.eid.unique()))
print(len(dfr_rec_a.uuid.unique()))

print(len(df_rec_a.eid.unique()))
print(len(df_rec_a.uuid.unique()))

missing_a=[i for i in df_rec_a.eid.unique().tolist() if i not in dfr_rec_a.eid.unique().tolist()]

missing_df=pd.DataFrame()
for m in missing_a:
    result = df_rec_a[(df_rec_a['eid']==m)&(df_rec_a['topic']=='unclassified')]
    missing_df=missing_df.append(result)

print(len(missing_df))
print(len(missing_df.uuid.unique()))
print(len(missing_df.eid.unique()))
print(len(missing_a))

dfr_rec_a = pd.concat([dfr_rec_a,missing_df],axis=0)
print(len(dfr_rec_a.eid.unique()))
print(len(dfr_rec_a.uuid.unique()))

missing_u=[i for i in df_rec_a.uuid.unique().tolist() if i not in dfr_rec_a.uuid.unique().tolist()]

missing_df=pd.DataFrame()
for m in missing_u:
    result = df_rec_a[(df_rec_a['uuid']==m)&(df_rec_a['topic']=='unclassified')]
    missing_df=missing_df.append(result)

dfr_rec_a = pd.concat([dfr_rec_a,missing_df],axis=0)
print(len(dfr_rec_a.eid.unique()))
print(len(dfr_rec_a.uuid.unique()))

len(dfr_rec_a)

dfr_rec_a=dfr_rec_a.merge(alias, on='eid')

dfr_rec_a.drop(columns = 'from_clean',inplace=True)
dfr_rec_a.columns = ['mid', 'address', 'user_rec', 'type', 'subject', 'body', 'date', 'bcc',
       'uuid_rec', 'eid_rec', 'topic', 'topic_h','user_sent','uuid_sent']

dfr_rec_a

print(len(dfr_rec_a.eid_rec.unique()))
print(len(dfr_rec_a.uuid_rec.unique()))
print(len(dfr_rec_a.uuid_sent.unique()))

dfr_rec_a.to_csv('D:/Capstone/dataset/database/N4J/internal_received_sent.csv')

dfr_rec_a2= pd.DataFrame()
for a in a_list:
    result = dfr[(dfr['eid']==a)]
    dfr_rec_a2=dfr_rec_a2.append(result)

So far we have created csvs for uploading to n4j using the following parameters:

1) from or sent by people within the 150 list - internal.csv

2) from within the 150 list, excludes includes 'unclassfied' emails, exclude 3rd parties - internal_received_sent.csv

dfr_rec_a2=dfr_rec_a2[dfr_rec_a2.type=='received']

dfr_rec_a2
dfr_rec_a2.to_csv('D:/Capstone/dataset/database/N4J/internal_received2.csv')

dfr_rec_a3= pd.DataFrame()
for a in a_list:
    result = df[(df['eid']==a)&(df['type']=='received')]
    dfr_rec_a3=dfr_rec_a3.append(result)

dfr_rec_a3
dfr_rec_a2.to_csv('D:/Capstone/dataset/database/N4J/internal_received3.csv')
