==> Notebooks/10. Position and employee Mapping.ipynb <==
import numpy as np 
import pandas as pd
import pickle

position = pd.read_csv('D:/Github/DSI8_Alpha/project-capstone/Enron Analysis/Resource/Employeestate.csv')
alias=pd.read_csv('D:/Capstone/dataset/database/N4J/alias.csv')
alias.drop(columns = 'Unnamed: 0',inplace=True)

class StrToBytes:
    def __init__(self, fileobj):
        self.fileobj = fileobj
    def read(self, size):
        return self.fileobj.read(size).encode()
    def readline(self, size=-1):
        return self.fileobj.readline(size).encode()

### Load the dictionary containing the dataset
with open('D:/Github/DSI8_Alpha/project-capstone/Enron Analysis/Resource/final_project_dataset.pkl', 'r') as data_file:
    data_dict = pickle.load(StrToBytes(data_file))

# dict to dataframe
df = pd.DataFrame.from_dict(data_dict, orient='index')
df.replace('NaN', np.nan, inplace = True)

df.drop(columns=[ 'to_messages', 'deferral_payments',
       'loan_advances', 'restricted_stock_deferred',
       'deferred_income', 'total_stock_value', 'expenses',
       'from_poi_to_this_person', 'exercised_stock_options', 'from_messages',
       'other', 'from_this_person_to_poi', 'poi', 'long_term_incentive',
       'shared_receipt_with_poi', 'restricted_stock','director_fees'],inplace=True)

df.dropna(subset=['email_address'],inplace=True)

df=df[np.isfinite(df['salary'])]

df.to_csv('D:/Capstone/dataset/support_files/employee_inc.csv')

alias['alias']=alias['from_clean'].str.extract(r"(.*?)@")
alias.head()

position=position.iloc[:,0:4]

position_alias = position.alias.unique().tolist()

print(len(position.alias.unique().tolist()))
print(len(alias.alias.unique().tolist()))
print(len(alias.uuid.unique().tolist()))

position = position.merge(alias, on ='alias')

print(len(position.uuid.unique().tolist()))
print(len(position.alias.unique().tolist()))
print(len(alias.uuid.unique().tolist()))

position.head()

position=position.fillna('Employee')

position.info()

position.columns


position.columns =(['alias', 'first_name', 'last_name', 'position', 'address', 'user',
       'eid', 'uuid'])

position.position.unique().tolist()

tochange=[ 'CEO', 'CRO', 'COO','Managing Director','Managing']
for i in range(len(position)):
    for c in tochange:
        if position.position[i] == c:
            position.position[i]= 'Csuite' 
tochange=[ 'Lawyer']
for i in range(len(position)):
    for c in tochange:
        if position.position[i] == c:
            position.position[i]= 'Employee' 

position.to_csv('D:/Capstone/dataset/support_files/employee_pos.csv')
